more FPE fixes
