+++ benchmarks/linear_programming/cuopt/run_mip.cpp
@@ -38,8 +38,6 @@
 #include <string>
 #include <vector>
 
-#include <fenv.h>
-
 #include "initial_problem_check.hpp"
 
 void merge_result_files(const std::string& out_dir,
@@ -293,8 +291,6 @@ void return_gpu_to_the_queue(std::unordered_map<pid_t, int>& pid_gpu_map,
 
 int main(int argc, char* argv[])
 {
-  feenableexcept(FE_DIVBYZERO | FE_INVALID);
-
   argparse::ArgumentParser program("solve_MIP");
 
   // Define all arguments with appropriate defaults and help messages
