+++ cpp/src/dual_simplex/presolve.cpp
@@ -13,8 +13,6 @@
 #include <dual_simplex/solve.hpp>
 #include <dual_simplex/tic_toc.hpp>
 
-#include <fenv.h>
-
 #include <cmath>
 #include <iostream>
 #include <numeric>
